+++ src/platform/macos/window.rs
@@ -225,7 +225,7 @@ impl WindowDelegate {
                 let state = &mut *(state as *mut DelegateState);
                 emit_resize_event(state);
                 let scale_factor = NSWindow::backingScaleFactor(*state.window) as f32;
-                emit_event(state, WindowEvent::HiDPIFactorChanged(scale_factor));
+                emit_event(state, WindowEvent::HiDpiFactorChanged(scale_factor));
             }
         }
 
@@ -234,7 +234,7 @@ impl WindowDelegate {
                 let state: *mut c_void = *this.get_ivar("winitState");
                 let state = &mut *(state as *mut DelegateState);
                 let scale_factor = NSWindow::backingScaleFactor(*state.window) as f32;
-                emit_event(state, WindowEvent::HiDPIFactorChanged(scale_factor));
+                emit_event(state, WindowEvent::HiDpiFactorChanged(scale_factor));
             }
         }
 
@@ -779,7 +779,7 @@ impl Window2 {
                 if attrs.always_on_top {
                     let _: () = msg_send![*window, setLevel:ffi::NSWindowLevel::NSFloatingWindowLevel];
                 }
-                
+
                 if let Some((x, y)) = pl_attrs.resize_increments {
                     if x >= 1 && y >= 1 {
                         let size = NSSize::new(x as _, y as _);